cherry pick from pr #2575 (#2586)

+++ paddleformers/transformers/qwen2/modeling.py
@@ -120,7 +120,7 @@ def apply_rotary_pos_emb(q, k, cos, sin, position_ids):
         sin = sin[position_ids].unsqueeze(2)  # [bs, seq_len, 1, dim]
     q_embed = (q * cos) + (rotate_half(q) * sin)
     k_embed = (k * cos) + (rotate_half(k) * sin)
-    return q_embed, k_embed
+    return q_embed.astype(q.dtype), k_embed.astype(q.dtype)
 
 
 class Qwen2Attention(nn.Layer):
